synchronizing with DiFX py3temp branch (step 1)
The branch TOP area has crap in it that can be deleted
eventually; and the doit script to rebuilt a nightly is
now saved in the PP area.  Updated difxcmp for the new
git repo.  Jan found that -f 0 failed if timerange wasn't
at least defined, thus a oneliner change to drivepolconvert.py

This branch now runs without errors on some of the golden
scans with a difx nightly built using the py3temp branch using
CASA 5.8.0 (the latest and last Py2 branch of CASA).