Reverted _PolConvert.cpp to git state.
Made adjustments in task_polconvert.py for the renaming of
the fringe files and adapted it to new calling sequence.
In the solve case, files to be plotted are also used for
conversion, so all of them need to be plotted...this is not
true in the CASA (ALMA) case, so there are now forks in the
logic flow for this.  Added some additional commentary.