Committing something that worked on 2017 sample.
Introduced debugNewIF to provide context to some of the error
messages (time management is somewhat confusing).  Initializes
the indices to -1 (not found) in DataIOSWIN::getNextMixedVis().
Removed the very verbose ERROR message and introduced convisok
to make it a oneline error.  This still leaves open what the
correct behavior should be with missing data.  Main bugs:

indices[...] > 0 replaced with indices[...] >= 0.

currentVis[i][k] needs to be zeroed when polproducts are missing
so that previously read data is not used in the next conversion.

Applied Jan's fix to call to getAmpRatio( ... im -> ii ... ).
(im is indexing the IFs to convert, ii is the actual IF value).

Collateral commits:

In DEBUG mode, do not pass verbosity to CalTable, otherwise one
is buried in amazing amounts of detail...this needs to be conditioned
on some other debug variable.

Code to print out elements of currentVis (wtf) is commented out
and will be removed in the next commit along with other cleanup.